// libs/package-management/src/project/project-types.ts
export type PackageName = string;

export interface PackageInfo {
  name: PackageName;
  dirpath: string;
  path: string;
  packageJson: PackageJson;
}